feat: change api to export factory function.
