+++ package.json
@@ -29,6 +29,7 @@
     "webpack": "^2.6.1"
   },
   "dependencies": {
+    "@types/memory-fs": "0.3.0",
     "memory-fs": "0.4.1",
     "resolve": "1.3.3"
   },
@@ -36,7 +37,6 @@
     "@types/debug": "0.0.29",
     "@types/extract-text-webpack-plugin": "2.1.0",
     "@types/jest": "19.2.4",
-    "@types/memory-fs": "0.3.0",
     "@types/resolve": "0.0.4",
     "@types/webpack": "2.2.15",
     "babel-core": "6.24.1",
